pyproject.toml
[tool.poetry]
name = "poigoe"
version = "0.0.0"
description = "Python package for realtime vocoding"
authors = ["34j <34j.95a2p@simplelogin.com>"]
license = "GNU General Public License v3"
readme = "README.md"
repository = "https://github.com/34j/poigoe"
documentation = "https://poigoe.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "poigoe", from = "src" },
]

[tool.poetry.scripts]
poigoe = 'poigoe.__main__:cli'

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/34j/poigoe/issues"
"Changelog" = "https://github.com/34j/poigoe/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.8"
sounddevice = "^0.4.6"
numpy = "^1.24.2"
soundfile = "^0.12.1"
matplotlib = "^3.7.0"
praat-parselmouth = "^0.4.3"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^3.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = ">=0.16"
sphinx = ">=4.0"
sphinx-rtd-theme = ">=1.0"

[tool.semantic_release]
branch = "main"
version_toml = "pyproject.toml:tool.poetry.version"
version_variable = "src/poigoe/__init__.py:__version__"
build_command = "pip install poetry && poetry build"

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=poigoe --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.isort]
profile = "black"
known_first_party = ["poigoe", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
